[what] 1. use svn external to link OpenCV2.1 2. Add Kalman Filter for TouchLib Filter
git-svn-id: http://imtophooksln.googlecode.com/svn/trunk@480 fa729b96-8d43-11de-b54f-137c5e29c83a

+++ ImProSln/TouchLibFilter/Touchlib/include/TouchScreenDevice.h
@@ -11,7 +11,7 @@ namespace touchlib
 	class  TouchScreenDevice 
 	{
 	public:
-		static ITouchScreen *getTouchScreen(float cw = 640.0, float ch = 480.0);
+		static ITouchScreen *getTouchScreen(int cw = 640, int ch = 480.0);
 		static void destroy();
 	private:
 		static CTouchScreen *tscreen;
